pyproject.toml
[project]
name = "trading-agent"
version = "0.1.0"
description = "Technical analyst agent built with LangGraph."
readme = "Readme.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-core",
    "langfuse",
    "python-dotenv",
    "langchain-ollama",
    "langchain-google-genai",
    "yfinance",
    "pandas",
    "numpy",
    "fastapi",
    "uvicorn",
    "google-genai",
    "openai",
    "requests",
    "beautifulsoup4",
    "pydantic-settings",
    "finnhub-python",
    "dash",
    "dash-bootstrap-components",
    "langchain-openai",
    "langchain-groq",
    "langchain-mistralai",
    "python-docx>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["technical_agent"]
